+++ modules/bridge/src/main/ts/ephox/bridge/components/dialog/Iframe.ts
@@ -5,18 +5,24 @@ import { FormComponentWithLabel, formComponentWithLabelFields, FormComponentWith
 
 export interface IframeSpec extends FormComponentWithLabelSpec {
   type: 'iframe';
+  border?: boolean;
   sandboxed?: boolean;
+  streamContent?: boolean;
   transparent?: boolean;
 }
 
 export interface Iframe extends FormComponentWithLabel {
   type: 'iframe';
+  border: boolean;
   sandboxed: boolean;
+  streamContent: boolean;
   transparent: boolean;
 }
 
 const iframeFields = formComponentWithLabelFields.concat([
+  FieldSchema.defaultedBoolean('border', false),
   FieldSchema.defaultedBoolean('sandboxed', true),
+  FieldSchema.defaultedBoolean('streamContent', false),
   FieldSchema.defaultedBoolean('transparent', true)
 ]);