+++ modules/tinymce/src/core/main/ts/content/InsertContentImpl.ts
@@ -1,5 +1,5 @@
 import { Arr, Optional, Type } from '@ephox/katamari';
-import { Replication, Class, Remove, SugarElement } from '@ephox/sugar';
+import { Remove, SugarElement } from '@ephox/sugar';
 
 import DOMUtils from '../api/dom/DOMUtils';
 import Editor from '../api/Editor';
@@ -181,16 +181,19 @@ const moveSelectionToMarker = (editor: Editor, marker: HTMLElement | null): void
   dom.remove(marker);
 
   if (parentBlock && dom.isEmpty(parentBlock)) {
+    const isCell = isTableCell(parentBlock);
+
     Remove.empty(SugarElement.fromDom(parentBlock));
 
     rng.setStart(parentBlock, 0);
     rng.setEnd(parentBlock, 0);
 
-    if (!isTableCell(parentBlock) && !isPartOfFragment(parentBlock) && (nextRng = findNextCaretRng(rng))) {
+    if (!isCell && !isPartOfFragment(parentBlock) && (nextRng = findNextCaretRng(rng))) {
       rng = nextRng;
       dom.remove(parentBlock);
     } else {
-      dom.add(parentBlock, dom.create('br', { 'data-mce-bogus': '1' }));
+      // TINY-9860: If parentBlock is a table cell, add a br without 'data-mce-bogus' attribute.
+      dom.add(parentBlock, dom.create('br', isCell ? {} : { 'data-mce-bogus': '1' }));
     }
   }
 
@@ -224,19 +227,6 @@ const findMarkerNode = (scope: AstNode): Optional<AstNode> => {
   }
 
   return Optional.none();
-};
-
-const preventSplittingSummary = (editor: Editor): void => {
-  Arr.each(Arr.from(editor.getBody().querySelectorAll('details')), (accordion) => {
-    const summaries = Arr.filter(Arr.from(accordion.children), (node) => node.nodeName === 'SUMMARY');
-    if (summaries.length > 1) {
-      Arr.each(summaries.slice(1), (summary) => {
-        const element = SugarElement.fromDom(summary);
-        Class.remove(element, 'mce-accordion-summary');
-        Replication.mutate(element, 'p');
-      });
-    }
-  });
 };
 
 export const insertHtmlAtCaret = (editor: Editor, value: string, details: InsertContentDetails): string => {
@@ -370,7 +360,6 @@ export const insertHtmlAtCaret = (editor: Editor, value: string, details: Insert
   moveSelectionToMarker(editor, dom.get('mce_marker'));
   unmarkFragmentElements(editor.getBody());
   trimBrsFromTableCell(dom, selection.getStart());
-  preventSplittingSummary(editor);
   TransparentElements.updateCaret(editor.schema, editor.getBody(), selection.getStart());
 
   return value;