+++ modules/tinymce/src/plugins/lists/test/ts/browser/ContentEditableFalseActionsTest.ts
@@ -146,4 +146,26 @@ ${listContent}
   it('TINY-9823: InsertUnorderedList command should ignore noneditable blocks', () => {
     testNonEditableBlocksIgnore('InsertUnorderedList');
   });
+
+  const testEditableBlockInsideNonEditableBlock = (command: 'InsertOrderedList' | 'InsertUnorderedList'): void => {
+    const editor = hook.editor();
+    editor.getBody().contentEditable = 'false';
+    const initialContent = '<div contenteditable="true"><p>a</p></div>';
+    const tag = command === 'InsertOrderedList' ? 'ol' : 'ul';
+    const expectedContent = `<div contenteditable="true">\n<${tag}>\n<li>a</li>\n</${tag}>\n</div>`;
+
+    editor.setContent(initialContent);
+    TinySelections.setSelection(editor, [ 0, 0, 0 ], 0, [ 0, 0, 0 ], 1);
+    editor.execCommand(command);
+    TinyAssertions.assertContent(editor, expectedContent);
+    editor.getBody().contentEditable = 'true';
+  };
+
+  it('TINY-10000: InsertOrderedList command should create a list in an editable block inside a non-editable block', () => {
+    testEditableBlockInsideNonEditableBlock('InsertOrderedList');
+  });
+
+  it('TINY-10000: InsertUnorderedList command should create a list in an editable block inside a non-editable block', () => {
+    testEditableBlockInsideNonEditableBlock('InsertUnorderedList');
+  });
 });