+++ modules/alloy/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 13.0.0 - 2023-07-12
+
+### Added
+- Added new `isBlocked` API to the `Blocking` behaviour.
+
+### Changed
+- `ModalDialogApis.getFooter` now returns an `Optional<AlloyComponent>` instead of `AlloyComponent`. #TINY-9996
+
 ### Fixed
 - Tab navigation did not forward focus over pseudo tab stops. #TINY-9815